[rename] signUpFormFields.ts -> fields.ts
